# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Crusher API
  description: |
    碎石记（Crusher）API 文档

    ## 认证
    大多数端点需要 Bearer Token 认证。Token 通过 Supabase Auth 获取。

    ## 错误响应
    所有错误响应遵循统一格式：
    ```json
    {
      "success": false,
      "error": "错误消息",
      "code": "错误码（可选）",
      "details": {}
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@crusher.com
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: https://crusher.com
    description: 生产环境
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: '错误消息'
        code:
          type: string
          example: 'ERROR_CODE'
        details:
          type: object
    ApiMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 5
        hasMore:
          type: boolean
          example: true

    # 用户相关
...
paths:
...
  /api/admin/credits:
    post:
      tags:
        - Admin
      summary: 创建积分调整请求（管理员）
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - amount
                - type
                - description
              properties:
                user_id:
                  type: string
                  format: uuid
                amount:
                  type: integer
                type:
                  type: string
                  enum: [manual_grant, admin_deduct]
                description:
                  type: string
      responses:
        '201':
          description: 积分调整请求创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                            format: uuid
                          status:
                            type: string
                            enum: [pending, approved, rejected]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/admin/credits/approve:
    post:
      tags:
        - Admin
      summary: 批准积分调整请求（管理员）
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: 积分调整请求批准成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          success:
                            type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/credits/pending:
    get:
      tags:
        - Admin
      summary: 获取待审批积分请求列表（管理员）
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功返回待审批积分请求列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            user_id:
                              type: string
                              format: uuid
                            amount:
                              type: integer
                            type:
                              type: string
                            description:
                              type: string
                            requested_by:
                              type: string
                            status:
                              type: string
                              enum: [pending, approved, rejected]
                            created_at:
                              type: string
                              format: date-time
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/credits/transactions:
    get:
      tags:
        - Admin
      summary: 获取积分交易历史（管理员）
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: 成功返回积分交易历史
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            user_id:
                              type: string
                              format: uuid
                            amount:
                              type: integer
                            balance_after:
                              type: integer
                            type:
                              type: string
                              enum: [payment, manual_grant, consumed, admin_deduct, refund]
                            description:
                              type: string
                              nullable: true
                            related_insight_id:
                              type: string
                              format: uuid
                              nullable: true
                            related_order_id:
                              type: string
                              format: uuid
                              nullable: true
                            operated_by:
                              type: string
                              format: uuid
                              nullable: true
                            created_at:
                              type: string
                              format: date-time
                      meta:
                        $ref: '#/components/schemas/ApiMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ==================== 管理后台 - 统计 API ====================
...
components:
  responses:
    Unauthorized:
      description: 未授权（未登录或 Token 无效）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: 'Unauthorized'
            code: 'UNAUTHORIZED'
    Forbidden:
      description: 禁止访问（权限不足）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: 'Forbidden'
            code: 'FORBIDDEN'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: 'Resource not found'
            code: 'NOT_FOUND'
    ValidationError:
      description: 请求参数验证失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: 'Validation failed'
            code: 'VALIDATION_ERROR'
...